chore: folder rename example app
