+++ examples/next-app/next.config.js
