+++ examples/next-app/package.json
