+++ examples/next-app/public/favicon.ico
