+++ examples/next-app/src/pages/_app.tsx
