+++ examples/next-app/src/pages/api/hello.ts
